+++ CHANGELOG.md
@@ -7,6 +7,11 @@ Date format: (YYYY-MM-DD)
 * Fixed: When a currency item is changed via the "/shopkeeper setCurrency" command, we close all currently open player shopkeeper UIs now so that the change is immediately in effect.
 * We no longer forcefully spawn Citizens shopkeeper NPCs when we reload their NPC data (e.g. when we apply a shopkeeper snapshot). Instead, the NPC is spawned based on its stored spawn state and location.
 * When placing shopkeepers with the shop creation item, we now ignore interactions with certain types of blocks that are relevant for navigation, such as doors, trap doors, fence doors, buttons, levers, etc. Other types of interactable blocks (work benches, chests, signs, etc.) are not ignored, because we want shopkeepers to be placeable against those (even if some plugins may use them for navigation purposes, e.g. command signs).
+* Added a new editor option to move shopkeepers.
+  * Config: Added `move-item` (default: `ENDER_PEARL`).
+  * Config: Moving of player shops can be disabled via the new `enable-moving-of-player-shops` setting (default: `true`).
+  * Some default messages related to the validation of spawn locations were slightly changed to better fit when we move shopkeepers.
+  * Limitation: Moving Citizens shopkeepers silently fails if the Citizens plugin is not enabled currently.
 
 **Debugging changes:**  
 * We log now when we unexpectedly encounter invalid currency items in the editors of the selling, buying, or book shopkeepers.
@@ -15,6 +20,18 @@ Date format: (YYYY-MM-DD)
 **Internal changes:**  
 * Refactors related to chat input handling.
 * Refactors related to shop creation, container selection, and shopkeeper placement.
+* Added InteractionInput for keeping track of interaction requests.
+* Added AbstractShopkeeper#teleport for moving a shopkeeper.
+* Added AbstractShopObject#move which moves the shop object to its intended spawn location based on its associated shopkeeper.
+
+**Message changes:**  
+* Changed `mob-cannot-spawn-on-peaceful-difficulty`.
+* Changed `restricted-area`.
+* Added `button-move`.
+* Added `button-move-lore`.
+* Added `click-new-shop-location`.
+* Added `shopkeeper-moved`.
+* Added `shopkeeper-move-aborted`.
 
 ## v2.15.1 (2022-03-27)
 ### Supported MC versions: 1.18.2, 1.18.1, 1.18, 1.17.1, 1.17, 1.16.5